Fix some review related code snippets
